# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Read-only API exposing fund snapshots, computed indicators, and chart data.\nAll numeric fields encoded as JSON strings to preserve full Stellar-stroop precision.",
        "title": "MTL Fund Statistics API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/api/v1/charts/balance-by-subfund": {
            "get": {
                "description": "Returns the EURMTL value of the 4 sub-fund accounts (MABIZ, MCITY, DEFI, BOSS) for a given date.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "charts"
                ],
                "summary": "Fund balance split by sub-fund",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Snapshot date (YYYY-MM-DD); defaults to latest",
                        "name": "date",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_api.BalanceBySubfundResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/charts/indicator-history": {
            "get": {
                "description": "Returns historical points for one or more indicator IDs over the requested range.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "charts"
                ],
                "summary": "Indicator time-series",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Comma-separated indicator IDs (e.g. 1,3,17,24,27)",
                        "name": "ids",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Range: 30d, 90d, 180d, 365d, or 'all' (default: 90d)",
                        "name": "range",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_api.IndicatorHistoryResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/indicators": {
            "get": {
                "description": "Returns indicators from the most recent stored snapshot. Optional `compare` adds period-over-period changes.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "indicators"
                ],
                "summary": "Latest indicators",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Comma-separated periods: any of 30d,90d,180d,365d, or 'all'",
                        "name": "compare",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal_api.IndicatorWithChanges"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/indicators/{date}": {
            "get": {
                "description": "Returns stored indicators for an exact snapshot date.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "indicators"
                ],
                "summary": "Indicators by date",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Snapshot date (YYYY-MM-DD)",
                        "name": "date",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_mtlprog_stat_internal_indicator.Indicator"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/snapshots": {
            "get": {
                "description": "Returns recent fund snapshots, newest first.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "snapshots"
                ],
                "summary": "List snapshots",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Maximum number of snapshots (default 30, max 365)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_mtlprog_stat_internal_snapshot.Snapshot"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/snapshots/latest": {
            "get": {
                "description": "Returns the most recent fund snapshot.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "snapshots"
                ],
                "summary": "Latest snapshot",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_mtlprog_stat_internal_snapshot.Snapshot"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/snapshots/{date}": {
            "get": {
                "description": "Returns the fund snapshot for an exact date.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "snapshots"
                ],
                "summary": "Snapshot by date",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Snapshot date (YYYY-MM-DD)",
                        "name": "date",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_mtlprog_stat_internal_snapshot.Snapshot"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_mtlprog_stat_internal_indicator.Indicator": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "unit": {
                    "type": "string"
                },
                "value": {
                    "type": "number"
                }
            }
        },
        "github_com_mtlprog_stat_internal_snapshot.Snapshot": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "entityId": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "snapshotDate": {
                    "type": "string"
                }
            }
        },
        "internal_api.BalanceBySubfundResponse": {
            "type": "object",
            "properties": {
                "date": {
                    "description": "YYYY-MM-DD",
                    "type": "string"
                },
                "slices": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_api.SubfundSlice"
                    }
                }
            }
        },
        "internal_api.HistoryPoint": {
            "type": "object",
            "properties": {
                "date": {
                    "description": "YYYY-MM-DD",
                    "type": "string"
                },
                "value": {
                    "type": "number"
                }
            }
        },
        "internal_api.IndicatorHistoryResponse": {
            "type": "object",
            "properties": {
                "series": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_api.IndicatorSeries"
                    }
                }
            }
        },
        "internal_api.IndicatorSeries": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "points": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_api.HistoryPoint"
                    }
                },
                "unit": {
                    "type": "string"
                }
            }
        },
        "internal_api.IndicatorWithChanges": {
            "type": "object",
            "properties": {
                "changes": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/internal_api.PeriodChange"
                    }
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "unit": {
                    "type": "string"
                },
                "value": {
                    "type": "number"
                }
            }
        },
        "internal_api.PeriodChange": {
            "type": "object",
            "properties": {
                "abs": {
                    "type": "number"
                },
                "pct": {
                    "type": "number"
                }
            }
        },
        "internal_api.SubfundSlice": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "value": {
                    "type": "number"
                }
            }
        }
    }
}