# internal/api/handler.go
package api

import (
	"encoding/json"
	"errors"
	"log/slog"
	"net/http"
	"strconv"
	"time"

	"github.com/mtlprog/stat/internal/snapshot"
)

// Handler provides HTTP endpoints for the statistics API.
type Handler struct {
	snapshots *snapshot.Service
}

// NewHandler creates a new API handler.
func NewHandler(snapshots *snapshot.Service) *Handler {
	return &Handler{snapshots: snapshots}
}

// GetLatestSnapshot handles GET /api/v1/snapshots/latest.
//
// @Summary      Latest snapshot
// @Description  Returns the most recent fund snapshot.
// @Tags         snapshots
// @Produce      json
// @Success      200  {object}  snapshot.Snapshot
// @Failure      404  {object}  map[string]string
// @Router       /api/v1/snapshots/latest [get]
func (h *Handler) GetLatestSnapshot(w http.ResponseWriter, r *http.Request) {
	s, err := h.snapshots.GetLatest(r.Context(), "mtlf")
	if err != nil {
		if errors.Is(err, snapshot.ErrNotFound) {
			writeError(w, http.StatusNotFound, "no snapshots found")
			return
		}
		slog.Error("failed to get latest snapshot", "error", err)
		writeError(w, http.StatusInternalServerError, "internal error")
		return
	}
	writeJSON(w, http.StatusOK, s)
}

// GetSnapshotByDate handles GET /api/v1/snapshots/{date}.
//
// @Summary      Snapshot by date
// @Description  Returns the fund snapshot for an exact date.
// @Tags         snapshots
// @Produce      json
// @Param        date  path  string  true  "Snapshot date (YYYY-MM-DD)"
// @Success      200  {object}  snapshot.Snapshot
// @Failure      400  {object}  map[string]string
// @Failure      404  {object}  map[string]string
// @Router       /api/v1/snapshots/{date} [get]
func (h *Handler) GetSnapshotByDate(w http.ResponseWriter, r *http.Request) {
	dateStr := r.PathValue("date")
	date, err := time.Parse("2006-01-02", dateStr)
	if err != nil {
		writeError(w, http.StatusBadRequest, "invalid date format, expected YYYY-MM-DD")
		return
	}

	s, err := h.snapshots.GetByDate(r.Context(), "mtlf", date)
	if err != nil {
		if errors.Is(err, snapshot.ErrNotFound) {
			writeError(w, http.StatusNotFound, "snapshot not found for date")
			return
		}
		slog.Error("failed to get snapshot by date", "date", dateStr, "error", err)
		writeError(w, http.StatusInternalServerError, "internal error")
		return
	}
	writeJSON(w, http.StatusOK, s)
}

// ListSnapshots handles GET /api/v1/snapshots.
//
// @Summary      List snapshots
// @Description  Returns recent fund snapshots, newest first.
// @Tags         snapshots
// @Produce      json
// @Param        limit  query  int  false  "Maximum number of snapshots (default 30, max 365)"
// @Success      200  {array}  snapshot.Snapshot
// @Router       /api/v1/snapshots [get]
func (h *Handler) ListSnapshots(w http.ResponseWriter, r *http.Request) {
	const maxLimit = 365
	limit := 30
	if l := r.URL.Query().Get("limit"); l != "" {
		if n, err := strconv.Atoi(l); err == nil && n > 0 {
			limit = min(n, maxLimit)
		}
	}

	snapshots, err := h.snapshots.List(r.Context(), "mtlf", limit)
	if err != nil {
		slog.Error("failed to list snapshots", "error", err)
		writeError(w, http.StatusInternalServerError, "internal error")
		return
	}
	writeJSON(w, http.StatusOK, snapshots)
}

func writeJSON(w http.ResponseWriter, status int, v any) {
	data, err := json.Marshal(v)
	if err != nil {
		slog.Error("failed to marshal JSON response", "error", err)
		http.Error(w, `{"error":"internal error"}`, http.StatusInternalServerError)
		return
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(status)
	if _, err := w.Write(data); err != nil {
		slog.Warn("failed to write HTTP response body", "error", err)
		return
	}
	_, _ = w.Write([]byte("\n"))
}

func writeError(w http.ResponseWriter, status int, msg string) {
	writeJSON(w, status, map[string]string{"error": msg})
}